// gen/ts-fetch/models/AdminSimSeedUsageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminSimSeedUsageResponse = {
    iccid?: string;
    usageDay?: string;
    visitedMccMnc?: string;
    uplinkKb?: number;
    downlinkKb?: number;
    totalKb?: number;
    seeded?: boolean;
};